[Feat] 리팩토링(코드 분리) & 달력에서 값 변경 후 메인뷰에 바로 반영
